requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12
sentence-transformers>=2.2.0
torch>=1.7.1
